+++ FrameController.lua
@@ -26,3 +26,22 @@ function FrameController:New(frame)
         fadeOnly = false,  -- Don't call Hide(), just set alpha to 0
         conditional = false,  -- Don't force Show() if frame is hidden
 
+        -- Buff frame anti-flicker (defer system)
+        deferFadeIn = false,
+        deferFadeOut = false,
+        deferReason = nil,
+    }
+
+    setmetatable(instance, self)
+    return instance
+end
+
+function FrameController:SetFadeOnly(value)
+    self.fadeOnly = value
+    return self
+end
+
+function FrameController:SetConditional(value)
+    self.conditional = value
+    return self
+end