+++ MinimapButton.lua
@@ -24,3 +24,9 @@ button:RegisterForClicks("LeftButtonUp", "RightButtonUp")
 button:RegisterForDrag("LeftButton")
 
 -- Button textures
+button:SetHighlightTexture("Interface\\Minimap\\UI-Minimap-ZoomButton-Highlight")
+
+local overlay = button:CreateTexture(nil, "OVERLAY")
+overlay:SetSize(53, 53)
+overlay:SetTexture("Interface\\Minimap\\MiniMap-TrackingBorder")
+overlay:SetPoint("TOPLEFT")