+++ Utils.lua
@@ -12,3 +12,6 @@ local Config = ZenHUD.Config  -- Already exported by Config.lua
 --------------------------------------------------------------------------------
 local Utils = {}
 
+function Utils.Print(msg, debugOnly)
+    if debugOnly and not Config:Get("debug") then return end
+    DEFAULT_CHAT_FRAME:AddMessage("|cFF66C2FF[ZenHUD]|r " .. msg)